Romdisk: add UART (test) commands and clean the organization of some routines
Added uartsnd and uartrcv commands to send and receive bytes to and from the UART (serial) driver respectively. These are only tests commands, they are not veryy useful currently but they let us test the driver at least.

The code has been re-organized a bit to have a file for str-related routines.