Date: fix in kernel & implementation on Zeal 8-bit Computer
Fix the get/setdate syscall in the system, fixed the corresponding syscall table/macros
Implement a getdate routine for Zeal 8-bit computer thanks to the I2C driver and the I2C RTC. (setdate not implemented yet)
The init program now presetns a `date` commands to show the current date and time. The routines have been refactored to be in the appropriate files.