VFS: fix a bug in the buffer size checker
A bug in the routine `zos_check_buffer_size` was allowing huge buffers
to be used with some syscalls, but this should not be allowed.
On the contrary, buffers of the size of a virtual page were not allowed,
but they should.
Added a comment in the public header files to explain which buffers
are allowed to be passed to syscalls.