zeal8bit: fix a bug in keyboard driver when UART is the standard output
When printing a whole buffer to the screen, the keyboard driver used to
end with a \0 character instead of a space (to remove duplicate char)
The UART monitor doesn't interpret \0 and simply discard them. This is also
the case of the video driver now. Thus, fix this by outputting a space instead,
to hide any duplicated character.