romdisk: re-implement `less` command to match its original behavior more
Since `init.bin` now has a cat command to print one or multiple files on the
standard output, the `less` command has been re-implemented to match a bit more
its original behavior (browse a file in an interactive manner).
At the moment, it is only possible to scroll up and down and exit, with `q`.