kernel: implement memory-persistent program when exec syscall invoked
It is now possible to keep the current program in memory when exec syscall
is invoked. The current program will get back the hand once the sub-program
finishes executing or if an error occurs. Thus, programs can now return a
value (main).
This feature is only available with kernels compiled with MMU support!