keyboard: define a blocking read and a non-blocking read mode
This commit brings definitions for a blocking and a non-blocking mode for the
keyboard. This will be useful in games that need to check keyboard events
without waiting for eys to be pressed and without re-implementing the driver(s).
This interface has been implemented on Zeal8bit target (cooked mode forced in blocking mode)