fs: add a new `HostFS` file system usable in emulated environments
This new file system is very convenient in an environment where Zeal 8-bit OS is being emulated. It lets the user share one of its host computer directory with Zeal 8-bit OS kernel.
Whenever the kernel will try to access a file or a directory on the HostFS disk,
the emulator will receive an I/O request that can be routed to the host computer
file system.
This feature is already implemented in Zeal 8-bit Computer emulator.